unpack-trees: add basic support for parallel checkout
This new interface allows us to enqueue some of the entries being
checked out to later uncompress, smudge, and write them in parallel. For
now, the parallel checkout machinery is enabled by default and there is
no user configuration, but run_parallel_checkout() just writes the
queued entries in sequence (without spawning additional workers). The
next patch will actually implement the parallelism and make it
configurable.

When there are path collisions among the entries being written (which
can happen e.g. with case-sensitive files in case-insensitive file
systems), the framework detects the problem and marks the item with
PC_ITEM_COLLIDED. Later, these items are sequentially fed to
checkout_entry() again. This is similar to the way the current code
deals with collisions, overwriting the previously checked out entries
with the subsequent ones. The only difference is that, since we no
longer create the files in the same order that they appear on index, we
are not able to determine which of the colliding entries will survive on
disk (for the current code, it is always the last one).

An alternative would be to simply skip the PC_ITEM_COLLIDED entries, and
not overwrite their colliding pairs. This idea seemed to work well in
simple tests, but because just one entry of each colliding group would
be actually written, the others would have null lstat() fields on the
index. This might not be a problem by itself, but it could cause
performance penalties for subsequent commands that need to refresh the
index: when the st_size value cached is 0, read-cache.c:ie_modified()
will go to the filesystem to see if the contents match. As mentioned in
the function:

    * Immediately after read-tree or update-index --cacheinfo,
    * the length field is zero, as we have never even read the
    * lstat(2) information once, and we cannot trust DATA_CHANGED
    * returned by ie_match_stat() which in turn was returned by
    * ce_match_stat_basic() to signal that the filesize of the
    * blob changed.  We have to actually go to the filesystem to
    * see if the contents match, and if so, should answer "unchanged".

So, if we have N entries in a colliding group and we decide to write and
lstat() only one of them, every subsequent git-status will have to read,
convert, and hash the written file N - 1 times, to check that the N - 1
unwritten entries are dirty. By checking out all colliding entries (like
the sequential code does), we only pay the overhead once, during
checkout.

Co-authored-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Co-authored-by: Jeff Hostetler <jeffhost@microsoft.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>