unpack-trees: add basic support for parallel checkout
This new interface allows us to enqueue some of the entries being
checked out to later call write_entry() for them in parallel. For now,
the parallel checkout machinery is enabled by default and there is no
user configuration, but run_parallel_checkout() just writes the queued
entries in sequence (without spawning additional workers). The next
patch will actually implement the parallelism and, later, we will make
it configurable.

When there are path collisions among the entries being written (which
can happen, for example, with case-sensitive files in case-insensitive
file systems), the parallel checkout code detects the problem and mark
the checkout_item with CI_RETRY. Later, these items are sequentially
feed to checkout_entry() again. This is how the sequential code deals
with collisions. The only difference is that, once we start writing the
entries in parallel, we will no longer be able to determine which of the
colliding entries will survive on disk (for the sequential algorithm, it
is always the last one). Also, there is no extra overhead when there
aren't collisions.

Note: we continue the loop of write_checkout_item() even if the previous
call returned an error. This is how checkout_entry() is called in
builtin/checkout.c:checkout_paths() and unpack-trees.c:check-updates().
In the case of fatal errors, die() aborts the loop.

Co-authored-by: Nguyễn Thái Ngọc Duy <pclouds@gmail.com>
Co-authored-by: Jeff Hostetler <jeffhost@microsoft.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>