compat/win32/pthread: add pthread_once()
Add pthread_once() emulation for Windows. This function allows
thread-safe one-time initializations in multithreaded code. It will be
used in the next commit to make hash_to_hex_algop() thread-safe.

This pthread_once() implementation comes from libav
(https://git.libav.org/?p=libav.git), commit b22693b ("w32pthreads: Add
pthread_once emulation", 2015-10-07). The code is licensed under
LGPLv2.1 which is compatible with GPLv2, as explained in section 3 of
LGPLv2.1. Only the section for support on Windows Vista+ has been
ported, as that's the minimum version required to build Git for Windows.
Also, the code was modified to (1) check and return error codes; and (2)
do not call InitOnceComplete() again after a successful initialization,
as that results in an error.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>