entry:write_entry(): try open() before other operations
In an upcoming patch we will introduce threads to checkout, which will
call write_entry() for multiple entries in parallel. In cases where
there might be path collisions (e.g. case-sensitive files on
case-insensitive file systems), the threads will compete for file
creation. However, as we don't want this behavior to be random, we will
record every write_entry() call that fails due to an open() error, and
sequentially let checkout_entry() take care of them, later. In this new
format, we don't want a thread to perform any work in write_entry()
before trying to create the file, as the work would be waisted in case
of a collision. Effectively, the file opening (with O_CREAT and
O_EXCL) will work as a lock, telling the thread that it either won the
race and shall continue or lost and shall return imediatelly, to
continue with another entry. With this in mind, reorder the operations
in this function to perform open() as the first step. Note that
streaming_write_entry() also had to be adjusted to receive a file
descriptor instead of the path string.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>