entry:write_entry(): categorize return codes
As mentioned in the previous patch, write_entry() will soon be called by
multiple threads in parallel. And we want to detect possible path
collisions when trying to create a file, so that we can better handle
them later. For this to work, we must be able to distinguish the types
of errors in write_entry(), so let's use a different return code for
each one of them. Note: we use defines instead of an enum because other
callers already expect write_entry()'s return type to be an int, and we
can't specify the size of an enum. We also only add categories for the
error types that we are interested in, the other errors are considered
generic errors.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>