compat/win32/pthread: emulate destructor callback for pthread_key
Our pthread_key_create() emulation for Windows currently ignores the
destructor callback, since TlsAlloc() doesn't take such a parameter.
This means that all users of the thread local storage (TLS) in our
codebase must remember to release the stored data before the thread
exits. Otherwise, there will be memory leaks on Windows. That's
unfortunate as we cannot use TLS, for example, to easily make functions
that return static buffers thread-safe. Furthermore, it's easy to
overlook the fact that the callback is ignored, which could lead to
erroneous uses of pthread_key_create() that leak memory on Windows.

We can overcome these problems using Windows' fibers API. Fibers are
finer execution units that must be manually scheduled. They inherit the
context of the thread that schedules them, but with extra specific data.
One of which is the fiber local storage (FLS). Unlike TlsAlloc(),
FlsAlloc() *does* take a destructor callback. And the interesting part
is that, as mentioned in [1], if there is no fiber switching, "FLS acts
exactly the same as thread local storage". Therefore, since we neither
use nor plan to use multiple fibers in the codebase, it should be safe
to replace TLS for FLS in our pthread_key emulation, allowing us to use
the destructor routine in Windows. With this change, threads will no
longer need to keep track of their thread local storage allocations to
later release the associated memory.

There is one caveat: the pthreads specification defines that "no
destructor functions shall be invoked by pthread_key_delete()", but
FlsFree() does call the destructor callback. So we cannot simply replace
the TLS functions with the FLS equivalents. Instead, to make the
emulation fully compatible we: (1) always register our own callback
function (so that we can control its behavior) and store the real
callback given by the caller in the key's struct. Then, (2) to pass
the real callback to our function, we wrap the FLS value in a struct
that contains a pointer to the destructor stored with the key. And
finally, (3) when pthread_key_delete() is called, we set the destructor
stored with that key to NULL so that our registered callback function
won't ever call it. Our callback is also responsible for releasing the
memory allocated for each FLS value struct.

[1]: https://docs.microsoft.com/en-us/windows/win32/procthread/fibers#fiber-local-storage

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>