compat/win32/pthreads: use destructor callback for pthread_key
Our pthread_key_create() emulation for Windows currently ignores the
destructor callback, since TLSAlloc() doesn't take such a parameter.
This means that all users of the thread local storage (TLS) in our
codebase must remember to release the stored data before the thread
exits. Otherwise, there will be memory leaks on Windows. That's
unfortunate as we need to keep track of all TLS allocations, and thus,
we cannot use TLS, for example, to easily make functions that return
static buffers thread-safe.

But we can overcome this problem using Windows' fibers API. Fibers are
finer execution units that must be manually scheduled. They inherit the
context of the thread that schedules them, but with extra specific data.
One of which is the fiber local storage (FLS). The interesting part is
that, as mentioned in [1], if there is no fiber switching, "FLS acts
exactly the same as thread local storage". And one of the key
differences in these two APIs is that FLSAlloc() *does* take a
destructor callback. Therefore, since we neither use nor plan to use
multiple fibers in the win32/pthreads, it should be safe to replace TLS
for FLS in our pthread_key emulation, allowing us to use the destructor
routine in Windows. With this change, threads will no longer need to
keep track of their thread local storage allocations to later release
the associated memory.

[1]: https://docs.microsoft.com/en-us/windows/win32/procthread/fibers#fiber-local-storage

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>