thread-utils: make pthread_key code functional with NO_PTHREADS
When NO_PTHREADS is defined, thread-utils.h provides a set of macros to
allow unconditional compilation for code files that use pthread
functions. But the macros regarding pthread_key don't currently provide
a functional emulation for single-threaded uses. Let's fix that so that
we can use these functions unconditionally in the next patch.

Note: this is the minimum emulation needed to make the code work without
pthreads. In particular, there is no error checking (e.g. if
pthread_key_destroy() is called for an uninitialized key). However, such
errors would already be present in the threaded code, so we should
still be able to detect a buggy caller.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>