thread-utils: make pthread_key code functional with NO_PTHREADS
When NO_PTHREADS is defined, thread-utils.h supplies a set of macros to
allow unconditional compilation for code files that use pthread
functions. But some macros regarding pthread_key don't currently provide
a functional emulation for single-threaded use. Let's fix that so that
we can call these functions unconditionally in the codebase. In
particular, this will be used in an upcoming patch to make
hash_to_hex_algop() thread-safe.

Note: this is the minimum emulation needed to make the code work without
pthreads. In particular, there is no error checking (e.g. if
pthread_key_destroy() is called for an uninitialized key).

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>