parallel-checkout: create leading dirs in workers
Allow the parallel workers to create the leading directories of the
entries being checked out, instead of pre-creating them in the main
process. This optimization should be more effective on file systems with
higher I/O latency that do take advantage of parallel I/O.

Part of the process of creating leading dirs is the removal of any
non-directory file that could be in the way. This is currently done
inside entry.c:create_directories(). However, if we were to move this to
the workers as well, we would risk removing a file just written by
another worker, which collided with the one currently being written. In
a worse scenario, we could remove the file right after a worker have
closed it but before it called stat(). To avoid these problems, let's
remove the non-directory files in the main process. There is no extra
cost since we already have the information on which path component (if
any) is not a directory.

Finally, to create the leading dirs in the workers, we could re-use
create_directories(). But, unlike the main process, the workers don't
have the stat() information cached. Thus, let's use
raceproof_create_file(), which will only stat() the path components
after a open() failure, saving us time when creating subsequent files in
the same directory.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>