parallel-checkout: skip checking the working tree on clone
When the checkout operation is part of a clone, there is no need to
stat() the entries' paths, as we know the working tree is empty. So
let's skip that, to save some time during clone.

Note: this optimization is not possible for sequential checkout (without
refactoring), because the same code that checks the working tree for
previously present files, is also used to detect and handle path
collisions during checkout. On parallel checkout, collisions are
detected later, by the workers themselves.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>