parallel-checkout: avoid duplicated lstat()'s in workers
Checkout workers must lstat() the path components of each entry before
writing to make sure that they are all real directories and not
symlinks. This check is already performed by the main process before
enqueueing the entries, but the second check is necessary in order to
avoid one specific path-collision problem. It can be seen in the
following scenario:

- First, the regular entry 'x/f' is processed:
  * We check that the directory part (in this case, 'x/') exists and
    that it is a real directory. If it doesn't exist, we create it.
  * Then we enqueue 'x/f' for parallel checkout.

- Next, the symlink 'X' gets processed. Let's suppose that, the path 'X'
  ends up colliding on disk with the previously created 'x' directory.
  So 'x' (dir) gets replaced by 'X' (symlink).

- When all entries have been either sequentially checked out or enqueued
  for parallel checkout, we start the workers. That's when 'x/f' will
  get written.

At this last step, if we don't check again that 'x' is a directory, and
if 'X' is a symlink to another directory, the worker will wrongly create
the file at 'X/f'.

We can, however, avoid this problem without the duplicated lstat()
calls, by postponing the creation of symlinks. Note that workers might
still encounter other types of path collision on the leading
directories. For example, the directory 'x' might be replaced by a
regular file 'X'. To be able to detect these collisions without the
additional lstat() calls we need to additionally look for ENOTDIR and
ENOENT when open() fails on workers.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>