parallel-checkout: avoid duplicated lstat()'s in workers
Checkout workers must lstat() the path components of each entry before
writing, to make sure that they are all real directories and not
symlinks. The same check is already performed by the main process before
enqueueing the entries, so we end up having some duplicated work. But
the second check is necessary to avoid a path-collision problem: if a
symlink X is created after an entry x/f was enqueued, and the file
system folding rules make x and X collide, the checkout worker could
wrongly create the file at X/f.

We can, however, avoid this problem without the duplicated lstat()
calls, by postponing the creation of symlinks. Note that workers might
still find other types of path collision when trying to checkout x/f
-- such as x being a regular file instead of a directory -- but these
cases are already handled when we check errno after an open() error.
This approach would not be possible with the symlink case, as open()
would succeed.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>