checkout--helper: only stat() entries if updating cache
When 'struct checkout's refresh_cache is false, there is no need for the
checkout workers to stat() the written entries nor to send this
information to the main process, since it won't be used. So let's add an
--[no]-stat option to checkout--helper to inform when stat() information
is really needed, and set it in parallel checkout according to
refresh_cache's value.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>