t1308-config-set: avoid false positives when using test-config
One test in t1308 expects test-config to fail with exit code 128 due to
a parsing error in the config machinery. But test-config might also exit
with 128 for any other reason that leads it to call die(). So the test
can potentially succeed for the wrong reason. To avoid false positives,
let's additionally check test-config's output in this test, to make sure
that the cause of the error is what we are expecting.

Moreover, the test was using the auxiliary function check_config which
optionally takes a string to compare against the test-config stdout.
Because this string is optional, there is a risk that future callers may
also check only the exit code and not the output. To avoid that, make
the string parameter of this function mandatory.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>