config: correctly read worktree configs in submodules
The config machinery is not able to read worktree configs from a
submodule in a process where the_repository represents the superproject
and extensions.worktreeConfig is not set there. Furthermore, when
extensions.worktreeConfig is set on the superproject, querying for a
worktree config in a submodule will, instead, return the value set at
the superproject.

The relevant code is in do_git_config_sequence(). Although it is
designed to act on an arbitrary repository, specified in the passed-in
`struct config_options`, it accidentally depends on the_repository in
two places:

- it reads the global variable `repository_format_worktree_config`,
  which is set based on the content of the_repository, to determine
  whether extensions.worktreeConfig is set.

- it uses the git_pathdup() helper to find the config.worktree file,
  instead of making a path using the passed-in repository.

Sever these dependencies and add a regression test. Also, to avoid
future misuses of `repository_format_worktree_config` like this one,
remove this global variable and store the config value on
`struct repository` itself.

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>